my_agent/requirements.txt
strands-agents>=1.8.0
strands-agents-tools>=0.2.7
mcp==1.13.1
python-dotenv